+++ Runner/src/GameStateStart.cpp
@@ -12,7 +12,7 @@ GameStateStart::GameStateStart(Game* game) : isDeletable(false)
 {
 	this->game = game;
 	sf::Vector2f pos = sf::Vector2f(this->game->window.getSize());
-	sf::FloatRect viewRect(0, 0, this->game->window.getSize().x, this->game->window.getSize().y);
+	sf::FloatRect viewRect(0, 0, static_cast<float>(this->game->window.getSize().x), static_cast<float>(this->game->window.getSize().y));
 	this->view.reset(viewRect);
 	pos *= 0.5f;
 	this->view.setCenter(pos);
@@ -25,8 +25,8 @@ void GameStateStart::setGameGUI()
 {
 	this->game->background.setPosition(this->game->window.mapPixelToCoords(sf::Vector2i(0, 0)));
 	this->game->background.setScale(
-		float(this->game->window.getSize().x) / float(this->game->background.getTexture()->getSize().x),
-		float(this->game->window.getSize().y) / float(this->game->background.getTexture()->getSize().y));
+		static_cast<float>(this->game->window.getSize().x) / static_cast<float>(this->game->background.getTexture()->getSize().x),
+		static_cast<float>(this->game->window.getSize().y) / static_cast<float>(this->game->background.getTexture()->getSize().y));
 
 	sf::Vector2f dim(GAME_MENU_BUTTON_WIDTH, GAME_MENU_BUTTON_HEIGHT);
 	int padding = 4;
@@ -39,22 +39,22 @@ void GameStateStart::setGameGUI()
 	};
 	this->guiElements.emplace("menu", std::make_shared<GuiList>(menu, false));
 	this->guiElements.at("menu")->setOrigin(GAME_MENU_BUTTON_WIDTH / 2, GAME_MENU_BUTTON_HEIGHT / 2);
-	this->guiElements.at("menu")->setPosition(sf::Vector2f(this->game->window.getSize().x, this->game->window.getSize().y) * 0.5f);
+	this->guiElements.at("menu")->setPosition(sf::Vector2f(static_cast<float>(this->game->window.getSize().x), static_cast<float>(this->game->window.getSize().y) * 0.5f));
 	this->guiElements.at("menu")->show();
 }
 
 void GameStateStart::resize(sf::Event& event)
 {
-	sf::Vector2f pos = sf::Vector2f(event.size.width, event.size.height);
-	sf::FloatRect viewRect(0, 0, event.size.width, event.size.height);
+	sf::Vector2f pos = sf::Vector2f(static_cast<float>(event.size.width), static_cast<float>(event.size.height));
+	sf::FloatRect viewRect(0, 0, static_cast<float>(event.size.width), static_cast<float>(event.size.height));
 
 	this->view.reset(viewRect);
 	this->view.setCenter(pos*0.5f);
 
 	this->game->background.setPosition(this->game->window.mapPixelToCoords(sf::Vector2i(0, 0)));
 	this->game->background.setScale(
-		float(event.size.width) / float(this->game->background.getTexture()->getSize().x),
-		float(event.size.height) / float(this->game->background.getTexture()->getSize().y));
+		static_cast<float>(event.size.width) / static_cast<float>(this->game->background.getTexture()->getSize().x),
+		static_cast<float>(event.size.height) / static_cast<float>(this->game->background.getTexture()->getSize().y));
 
 	this->guiElements.at("menu")->setPosition(pos * 0.5f);
 	this->guiElements.at("menu")->show();
@@ -102,15 +102,15 @@ void GameStateStart::handleInput() {
 
 		case sf::Event::Resized:
 		{
-			this->view.setSize(event.size.width, event.size.height);
+			this->view.setSize(static_cast<float>(event.size.width), static_cast<float>(event.size.height));
 			this->game->background.setPosition(this->game->window.mapPixelToCoords(sf::Vector2i(0, 0),this->view));
-			sf::Vector2f pos = sf::Vector2f(event.size.width, event.size.height);
+			sf::Vector2f pos = sf::Vector2f(static_cast<float>(event.size.width), static_cast<float>(event.size.height));
 			pos *= 0.5f;
 			pos = this->game->window.mapPixelToCoords(sf::Vector2i(pos), this->view);
 			this->guiElements.at("menu")->setPosition(pos);
 			this->game->background.setScale(
-				float(event.size.width) / float(this->game->background.getTexture()->getSize().x),
-				float(event.size.height) / float(this->game->background.getTexture()->getSize().y));
+				static_cast<float>(event.size.width) / static_cast<float>(this->game->background.getTexture()->getSize().x),
+				static_cast<float>(event.size.height) / static_cast<float>(this->game->background.getTexture()->getSize().y));
 			break;
 		}
 		case sf::Event::MouseButtonPressed:
